Small clenaup in the python scripts to make sure all tests get their Cleanup method called. Review URL: https://chromereviews.googleplex.com/3534013
git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart@108 260f80e4-7a28-3924-810f-c04153c831b5

+++ tools/run.py
@@ -285,8 +285,9 @@ class BrowserArchitecture(Architecture):
     return 1
 
   def Cleanup(self):
-    shutil.rmtree(self.temp_dir)
-    self.temp_dir = None
+    if self.temp_dir:
+      shutil.rmtree(self.temp_dir)
+      self.temp_dir = None
 
 
 class ChromiumArchitecture(BrowserArchitecture):
